+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 3.1.3
 * Disable collectd-java plugin installation
+* Add InSpec checks to ensure java does not use OpenJDK
 * Update Hiera configuration for Author, Publish & Consolidated to support custom configuration Parameters for waiting until login page is ready. shinesolutions/aem-aws-stack-builder#184
 * Add configuration parameters for Chaos Monkey to configure component termination rule #122
 * Improved logging for taking live/offline snapshot